pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mpci"
dynamic = ["version", "readme"]
description = "Multi-Photon Calcium Imaging"
authors = [
    {name = "IBL Staff"},
]
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    # "plane2brain",
    "ibllib",
    "numpy",
    "pandas",
    "ONE-api",
    "iblutil",
    "packaging",
    "scipy",
    "scikit-image",
    "opencv-python-headless",
    "tqdm",
    "iblatlas"
]

[project.urls]
Homepage = "https://github.com/int-brain-lab/mpci"
Documentation = "https://one.internationalbrainlab.org"
Repository = "https://github.com/int-brain-lab/mpci.git"
Issues = "https://github.com/int-brain-lab/mpci/issues"

[tool.setuptools.dynamic]
version = {attr = "one.__version__"}
readme = {file = ["README.md"], content-type = "text/markdown"}

[tool.setuptools.packages.find]
include = ["mpci*"]
exclude = ["mpci.tests*"]

[tool.ruff]
line-length = 99
extend-exclude = ["docs"]

[tool.ruff.lint]
select = ["D", "E", "F", "W"]
ignore = ["E266", "D401", "D105", "D107", "D102", "D417", "D101", "D404", "D100", "D301", "D103", "D403"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
